requirements.txt
langchain
langchain-community
langchain-anthropic
langchain-groq
langchain-text-splitters
langgraph
langgraph-checkpoint-postgres
psycopg[binary]
psycopg2-binary
langfuse
qdrant-client
langchain-qdrant
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
groq
fastembed
pypdf
cryptography>=3.1
python-multipart
requests
beautifulsoup4
pandas
fpdf2
pydantic-settings>=2.0
pytesseract>=0.3.10
pdf2image>=1.17.0
